% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plots.R
\name{AC_filter_plot}
\alias{AC_filter_plot}
\title{Assignment-Control Filter Plot}
\usage{
AC_filter_plot(data, match, title = "")
}
\arguments{
\item{data}{a data.frame or tibble with t, prop, and prog}

\item{match}{an optmatch object with a 1:1 matching of subjects}

\item{title}{a title for the plot}
}
\value{

}
\description{
Shows an AC plot of a matched dataset, with matched individuals opaque and
unmatched individuals translucent
}
