% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAC_plots.R
\name{AC_plot}
\alias{AC_plot}
\title{Assignment-Control Plot}
\usage{
AC_plot(data, title = "", opaque_class = 1)
}
\arguments{
\item{data}{a data.frame or tibble with t, prop, and prog}

\item{title}{a title for the plot}

\item{opaque_class}{1, 0 or "none".  Specifies which class should be opaque
and which should be translucent}
}
\value{
a ggplot object
}
\description{
Makes an AC plot of the data, with treated individuals in red and opaque,
control individuals translucent blue.  Assumes propensity and prognostic
scores are already in the data as \code{prop} and \code{prog}
}
