% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plots.R
\name{CR_filter_plot}
\alias{CR_filter_plot}
\title{Control-Randomization Filter Plot}
\usage{
CR_filter_plot(data, match, title = "")
}
\arguments{
\item{data}{a data.frame or tibble with t, prop, and prog}

\item{match}{an optmatch object with a 1:1 matching of subjects}

\item{title}{a title for the plot}
}
\value{

}
\description{
Shows a CR plot of a matched dataset, with matched individuals opaque and
unmatched individuals translucent
}
