% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate simulated data}
\usage{
generate_data(
  N = 2000,
  p = 10,
  true_mu = "X1/3-3",
  rho = 0,
  sigma = 1,
  tau = 1
)
}
\arguments{
\item{N}{numeric, sample size}

\item{p}{numeric, number of features}

\item{true_mu}{string formula giving true propensity score linear model}

\item{rho}{numeric between 0 and 1.  0 => prog orthogonal to prop, 1=> prog
|| prop}

\item{sigma}{numeric noise to be added to y. y += sigma*rnorm(0,1)}

\item{tau}{numeric additive treatment effect}
}
\value{
data.frame of covariates, y, t, and true prop (logit scale) and prog
}
\description{
Data set contains measured covariates X, outcome Y, treatment
assignment t, and logit(propensity), mu.

covariate data ~ normal(0,1); mu = true_mu; t ~ binom(p = 1 / (1 +
exp(-mu))); y ~ rho * X1 + sqrt(1-rho^2) * X2 + tau * t + epsilon epsilon ~
normal(0, 1)
}
