% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data_IV_xSITA}
\alias{generate_data_IV_xSITA}
\title{Generate Simulated IV data - with a SITA violation}
\usage{
generate_data_IV_xSITA(
  N = 2000,
  p = 10,
  true_mu = "X1/2 + Z/2 - 3.25",
  rho = 0,
  rho_z = 0.2,
  sigma = 1,
  tau = 1
)
}
\arguments{
\item{N}{numeric, sample size}

\item{p}{numeric, number of features}

\item{true_mu}{string formula giving true propensity score linear model}

\item{rho}{numeric between 0 and 1.  0 => prog orthogonal to prop, 1=> prog
|| prop}

\item{rho_z}{numeric between 0 and 1.  correllation between Z and X1}

\item{sigma}{numeric noise to be added to y. y += sigma*rnorm(0,1)}

\item{tau}{numeric additive treatment effect}
}
\value{
data.frame of covariates, y, t, and true prop (logit scale) and prog
}
\description{
Data set contains measured covariates X, outcome Y, treatment
assignment t, and logit(propensity), mu.

covariate data ~ normal(0,1); mu = true_mu; t ~ binom(p = 1 / (1 +
exp(-mu))); y ~ rho * X1 + sqrt(1-rho^2) * X2 + tau * t + epsilon epsilon ~
normal(0, 1); Z instrument correllated with X1 by rho_z
}
