% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_rd_data}
\alias{generate_rd_data}
\title{Generate simulated data for RD design}
\usage{
generate_rd_data(
  N = 1000,
  p = 10,
  c = qnorm(0.75),
  rho = 0,
  sigma = 1,
  tau = 1
)
}
\arguments{
\item{N}{numeric, sample size}

\item{p}{numeric, number of features}

\item{c}{numeric, cutoff for forcing variable}

\item{rho}{numeric between 0 and 1.  0 => prog orthogonal to prop, 1=> prog
|| prop}

\item{sigma}{numeric noise to be added to y. y += sigma*rnorm(0,1)}

\item{tau}{numeric additive treatment effect}
}
\value{
data.frame of covariates, y, t, and mu
}
\description{
Data set contains measured covariates X, outcome Y, treatment
assignment t, and logit(propensity), mu.

covariate data ~ normal(0,1);
t = 0 if X1 < c, 1 otherwise;
y ~ rho * X1 + sqrt(1-rho^2) * X2 + tau * t + epsilon;
epsilon ~ normal(0, 1)
}
