% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_xSITA_data}
\alias{generate_xSITA_data}
\title{Generate simulated data with a SITA violation}
\usage{
generate_xSITA_data(
  N = 2000,
  p = 10,
  true_mu = "X1/3-3 + nu * U",
  rho = 0.1,
  nu = 0.2,
  sigma = 1,
  tau = 1
)
}
\arguments{
\item{N}{numeric, sample size}

\item{p}{numeric, number of features}

\item{true_mu}{string formula giving true propensity score linear model}

\item{rho}{numeric between 0 and 1.  0 => prog orthogonal to prop, 1=> prog
|| prop}

\item{nu}{coefficient of unmeasured confounder in propensity and prognosis}

\item{sigma}{numeric noise to be added to y. y += sigma*rnorm(0,1)}

\item{tau}{numeric additive treatment effect}
}
\value{
data.frame of covariates, y, t, and true prop (logit scale) and prog
}
\description{
Data set contains measured covariates X, unmeasured covariate U,
outcome Y, treatment assignment t, and logit(propensity), mu.

covariate data ~ normal(0,1);
mu = true_mu;
t ~ binom(p = 1 / (1 + exp(-mu)));
y ~ rho * X1 + sqrt(1-rho^2) * X2 + 0.2 * U + t + epsilon
epsilon ~ normal(0, 1)
}
