% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_histograms.R
\name{propensity_histogram}
\alias{propensity_histogram}
\title{Propensity Score Histogram}
\usage{
propensity_histogram(data)
}
\arguments{
\item{data}{data a data.frame or tibble with t and prop}
}
\value{

}
\description{
Propensity Score Histogram
}
