% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plots.R
\name{RA_filter_plot}
\alias{RA_filter_plot}
\title{Randomization-Assignment Filter Plot}
\usage{
RA_filter_plot(data, match, k = 1, title = "")
}
\arguments{
\item{data}{a data.frame or tibble with t, prop, and IV}

\item{match}{an optmatch object with a 1:1 matching of subjects}

\item{title}{a title for the plot}
}
\value{

}
\description{
Shows an RA plot of a matched dataset, with matched individuals opaque and
unmatched individuals translucent
}
