% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAC_plots.R
\name{RA_plot}
\alias{RA_plot}
\title{Randomization-Assignment plot}
\usage{
RA_plot(data, title = "", opaque_class = 1)
}
\arguments{
\item{data}{a data.frame or tibble with t, prop, and IV}

\item{title}{a title for the plot}

\item{opaque_class}{1, 0 or "none".  Specifies which class should be opaque
and which should be translucent}
}
\value{
a ggplot object
}
\description{
Makes an RA plot of the data, with treated individuals in red and opaque,
control individuals translucent blue.  Assumes propensity scores and IV are
already in the data as \code{prop} and \code{IV}
}
