% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plots.R
\name{RC_filter_plot}
\alias{RC_filter_plot}
\title{Randomization-Control Filter Plot}
\usage{
RC_filter_plot(data, match, title = "")
}
\arguments{
\item{data}{a data.frame or tibble with t, prog, and IV}

\item{match}{an optmatch object with a 1:1 matching of subjects}

\item{title}{a title for the plot}
}
\value{

}
\description{
Shows an RC plot of a matched dataset, with matched individuals opaque and
unmatched individuals translucent
}
