% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_plots.R
\name{RC_match_plot}
\alias{RC_match_plot}
\title{Randomization-Control plot with matching overlayed}
\usage{
RC_match_plot(data, match, title = "")
}
\arguments{
\item{data}{a data.frame or tibble with t, prog, and IV}

\item{match}{an optmatch object with a 1:1 matching of subjects}

\item{title}{a title for the plot}
}
\value{

}
\description{
Randomization-Control plot with matching overlayed
}
