% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAC_plots.R
\name{RC_plot}
\alias{RC_plot}
\title{Randomization-Control plot}
\usage{
RC_plot(data, title = "")
}
\arguments{
\item{data}{a data.frame or tibble with t, prog, and IV}

\item{title}{a title for the plot}
}
\value{
a ggplot object
}
\description{
Makes an RC plot of the data, with treated individuals in red and opaque,
control individuals translucent blue.  Assumes prognostic
scores and IV are already in the data as \code{prog} and \code{IV}
}
